requirements.txt
beautifulsoup4==4.12.2
requests==2.28.1
selenium==4.8.2
webdriver-manager==4.0.2
python-dotenv==1.0.0
